refactor: #101 update client file name
